student-management-system/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.4
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
pandas==2.2.2
openpyxl==3.1.2
pycryptodome==3.19.0 